+++ tests/test_openh264.py
@@ -72,7 +72,7 @@ def test_openh264_sendonly_recvonly(setup):
 @pytest.mark.parametrize(
     "video_codec_type,expected_implementation",
     [
-        ("H264", "SimulcastEncoderAdapter (OpenH264, OpenH264, OpenH264)"),
+        ("H264", "OpenH264"),
     ],
 )
 def test_openh264_simulcast(setup, video_codec_type, expected_implementation):
@@ -122,6 +122,7 @@ def test_openh264_simulcast(setup, video_codec_type, expected_implementation):
 
     for i, rtp_stat in enumerate(sorted_stats):
         assert rtp_stat["rid"] == f"r{i}"
-        assert rtp_stat["encoderImplementation"] == expected_implementation
+        assert "SimulcastEncoderAdapter" in rtp_stat["encoderImplementation"]
+        assert expected_implementation in rtp_stat["encoderImplementation"]
         assert rtp_stat["bytesSent"] > 0
         assert rtp_stat["packetsSent"] > 0